+++ config/test.exs
@@ -2,7 +2,7 @@ use Mix.Config
 
 # We don't run a server during test. If one is required,
 # you can enable the server option below.
-config :janus_ws_example, JanusWsExampleWeb.Endpoint,
+config :janus_ws_example, Web.Endpoint,
   http: [port: 4002],
   server: false
