--- lib/janus_ws_example_web/controllers/page_controller.ex
@@ -1,7 +1,0 @@
-defmodule JanusWsExampleWeb.PageController do
-  use JanusWsExampleWeb, :controller
-
-  def index(conn, _params) do
-    render(conn, "index.html")
-  end
-end
