--- lib/janus_ws_example_web/router.ex
@@ -1,26 +1,0 @@
-defmodule JanusWsExampleWeb.Router do
-  use JanusWsExampleWeb, :router
-
-  pipeline :browser do
-    plug :accepts, ["html"]
-    plug :fetch_session
-    plug :fetch_flash
-    plug :protect_from_forgery
-    plug :put_secure_browser_headers
-  end
-
-  pipeline :api do
-    plug :accepts, ["json"]
-  end
-
-  scope "/", JanusWsExampleWeb do
-    pipe_through :browser
-
-    get "/", PageController, :index
-  end
-
-  # Other scopes may use custom stacks.
-  # scope "/api", JanusWsExampleWeb do
-  #   pipe_through :api
-  # end
-end
