--- lib/janus_ws_example_web/views/error_view.ex
@@ -1,16 +1,0 @@
-defmodule JanusWsExampleWeb.ErrorView do
-  use JanusWsExampleWeb, :view
-
-  # If you want to customize a particular status code
-  # for a certain format, you may uncomment below.
-  # def render("500.html", _assigns) do
-  #   "Internal Server Error"
-  # end
-
-  # By default, Phoenix returns the status message from
-  # the template name. For example, "404.html" becomes
-  # "Not Found".
-  def template_not_found(template, _assigns) do
-    Phoenix.Controller.status_message_from_template(template)
-  end
-end
