--- lib/janus_ws_example_web/views/layout_view.ex
@@ -1,3 +1,0 @@
-defmodule JanusWsExampleWeb.LayoutView do
-  use JanusWsExampleWeb, :view
-end
