--- lib/janus_ws_example_web/views/page_view.ex
@@ -1,3 +1,0 @@
-defmodule JanusWsExampleWeb.PageView do
-  use JanusWsExampleWeb, :view
-end
