+++ lib/web/controllers/page_controller.ex
@@ -1,0 +1,7 @@
+defmodule Web.PageController do
+  use Web, :controller
+
+  def index(conn, _params) do
+    render(conn, "index.html")
+  end
+end
