+++ lib/web/router.ex
@@ -1,0 +1,18 @@
+defmodule Web.Router do
+  use Web, :router
+
+  pipeline :browser do
+    plug :accepts, ["html"]
+    plug :put_secure_browser_headers
+  end
+
+  pipeline :api do
+    plug :accepts, ["json"]
+  end
+
+  scope "/", Web do
+    pipe_through :browser
+
+    get "/", PageController, :index
+  end
+end
