+++ lib/web/views/layout_view.ex
@@ -1,0 +1,3 @@
+defmodule Web.LayoutView do
+  use Web, :view
+end
