+++ lib/web/views/page_view.ex
@@ -1,0 +1,3 @@
+defmodule Web.PageView do
+  use Web, :view
+end
