+++ mix.exs
@@ -7,7 +7,7 @@ defmodule JanusWsExample.MixProject do
       version: "0.1.0",
       elixir: "~> 1.7",
       elixirc_paths: elixirc_paths(Mix.env()),
-      compilers: [:phoenix, :gettext] ++ Mix.compilers(),
+      compilers: [:phoenix] ++ Mix.compilers(),
       start_permanent: Mix.env() == :prod,
       deps: deps()
     ]
@@ -18,7 +18,7 @@ defmodule JanusWsExample.MixProject do
   # Type `mix help compile.app` for more information.
   def application do
     [
-      mod: {JanusWsExample.Application, []},
+      mod: {JanusEx.Application, []},
       extra_applications: [:logger, :runtime_tools]
     ]
   end
@@ -36,7 +36,6 @@ defmodule JanusWsExample.MixProject do
       {:phoenix_pubsub, "~> 1.1"},
       {:phoenix_html, "~> 2.11"},
       {:phoenix_live_reload, "~> 1.2", only: :dev},
-      {:gettext, "~> 0.11"},
       {:jason, "~> 1.0"},
       {:plug_cowboy, "~> 2.0"},
       {:janus_ws, "~> 0.1.0"},
