--- test/janus_ws_example_web/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule JanusWsExampleWeb.PageControllerTest do
-  use JanusWsExampleWeb.ConnCase
-
-  test "GET /", %{conn: conn} do
-    conn = get(conn, "/")
-    assert html_response(conn, 200) =~ "Welcome to Phoenix!"
-  end
-end
