--- test/janus_ws_example_web/views/error_view_test.exs
@@ -1,15 +1,0 @@
-defmodule JanusWsExampleWeb.ErrorViewTest do
-  use JanusWsExampleWeb.ConnCase, async: true
-
-  # Bring render/3 and render_to_string/3 for testing custom views
-  import Phoenix.View
-
-  test "renders 404.html" do
-    assert render_to_string(JanusWsExampleWeb.ErrorView, "404.html", []) == "Not Found"
-  end
-
-  test "renders 500.html" do
-    assert render_to_string(JanusWsExampleWeb.ErrorView, "500.html", []) ==
-             "Internal Server Error"
-  end
-end
