--- test/janus_ws_example_web/views/layout_view_test.exs
@@ -1,3 +1,0 @@
-defmodule JanusWsExampleWeb.LayoutViewTest do
-  use JanusWsExampleWeb.ConnCase, async: true
-end
