--- test/janus_ws_example_web/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule JanusWsExampleWeb.PageViewTest do
-  use JanusWsExampleWeb.ConnCase, async: true
-end
