+++ test/web/controllers/page_controller_test.exs
@@ -1,0 +1,8 @@
+defmodule Web.PageControllerTest do
+  use Web.ConnCase
+
+  test "GET /", %{conn: conn} do
+    conn = get(conn, "/")
+    assert html_response(conn, 200) =~ "Janus"
+  end
+end
