+++ test/web/views/layout_view_test.exs
@@ -1,0 +1,3 @@
+defmodule Web.LayoutViewTest do
+  use Web.ConnCase, async: true
+end
