+++ test/web/views/page_view_test.exs
@@ -1,0 +1,3 @@
+defmodule Web.PageViewTest do
+  use Web.ConnCase, async: true
+end
